Chitralochana_app/requirements.txt
Flask==0.11.1
Flask-Bootstrap==3.3.7.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.0
Flask-Moment==0.5.1
Flask-SQLAlchemy==2.1
Flask-Script==2.0.5
Flask-WTF==0.13.1
Jinja2==2.8
Mako==1.0.5
MarkupSafe==0.23
SQLAlchemy==1.1.3
WTForms==2.1
Werkzeug==0.11.11
alembic==0.8.8
blinker==1.4
itsdangerous==0.24
Flask-Babel==0.11.1
Flask-Login==0.4.0
flask-mongoengine==0.8
Flask-SocketIO==2.7.1
mongoengine==0.10.6
nltk==3.2.1
numpy==1.11.2+mkl
pymongo==3.3.1
python-socketio==1.6.0
sqlalchemy-migrate==0.10.0
tweepy=3.5.0